requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.22
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.25
email-validator==2.1.0
